backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
web3==6.11.3
python-dotenv==1.0.0
requests==2.31.0
eth-account==0.10.0
py-solc-x==2.0.2
